Changes as per PR comments
