+++ extensions/azurePublishNew/src/recoilModel/atoms/handOffToAdminState.ts
@@ -6,11 +6,11 @@ import { atom } from 'recoil';
 import { ResourcesItem } from '../../types';
 
 export const enabledHandOffResourcesState = atom<ResourcesItem[] | undefined>({
-  key: 'hor_optional_resources',
+  key: 'handoff_optional_resources',
   default: undefined,
 });
 
 export const requiredHandOffResourcesState = atom<ResourcesItem[]>({
-  key: 'hor_required_resources',
+  key: 'handoff_required_resources',
   default: [],
 });
